pack data into 32-bit, to mirror wasm32

+++ src/binparse.odin
@@ -30,25 +30,25 @@ init_parser :: proc(total_size: u32) -> Parser {
 	return p
 }
 
-setup_pid :: proc(process_id: u32) -> int {
+setup_pid :: proc(process_id: u32) -> i32 {
 	p_idx, ok := vh_find(&process_map, process_id)
 	if !ok {
 		append(&processes, init_process(process_id))
-		p_idx = len(processes) - 1
+		p_idx = i32(len(processes) - 1)
 		vh_insert(&process_map, process_id, p_idx)
 	}
 
 	return p_idx
 }
 
-setup_tid :: proc(p_idx: int, thread_id: u32) -> int {
+setup_tid :: proc(p_idx: i32, thread_id: u32) -> i32 {
 	t_idx, ok := vh_find(&processes[p_idx].thread_map, thread_id)
 	if !ok {
 		threads := &processes[p_idx].threads
 
 		append(threads, init_thread(thread_id))
 
-		t_idx = len(threads) - 1
+		t_idx = i32(len(threads) - 1)
 		thread_map := &processes[p_idx].thread_map
 		vh_insert(thread_map, thread_id, t_idx)
 	}
@@ -220,7 +220,7 @@ load_binary_chunk :: proc(chunk: []u8) {
 	return
 }
 
-bin_push_event :: proc(process_id, thread_id: u32, event: ^Event) -> (int, int, int) {
+bin_push_event :: proc(process_id, thread_id: u32, event: ^Event) -> (i32, i32, i32) {
 	p_idx := setup_pid(process_id)
 	t_idx := setup_tid(p_idx, thread_id)
 
@@ -251,7 +251,7 @@ bin_push_event :: proc(process_id, thread_id: u32, event: ^Event) -> (int, int, 
 	t.current_depth += 1
 	append_event(&depth.bs_events, event^)
 
-	return p_idx, t_idx, len(depth.bs_events)-1
+	return p_idx, t_idx, i32(len(depth.bs_events)-1)
 }
 
 bin_process_events :: proc() {
